+++ apps/docs/src/app/_components/ComposingCodeExample/ComposingCodeExample.module.d.scss.ts
@@ -1,0 +1,7 @@
+declare const classNames: {
+  readonly example: "example";
+  readonly editorContainer: "editorContainer";
+  readonly preview: "preview";
+  readonly codeEditor: "codeEditor";
+};
+export default classNames;